+++ src/app/(landing-routes)/blog/latest/page.tsx
@@ -1,6 +1,6 @@
 import BlogCard from "~/components/common/blogCard/BlogCard";
 import CustomButton from "~/components/common/common-button/common-button";
-import { articlesData } from "./data/dummy-article-data";
+import { articlesData } from "../data/dummy-article-data";
 
 const LatestArticlesPage = () => {
   const isTesting: boolean = process.env.NEXT_PUBLIC_TESTING === "true";
